Feat

+++ src/main/java/com/serjn/online/DTOs/ClientDto.java
@@ -3,12 +3,14 @@ package com.serjn.online.DTOs;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 
+import java.math.BigDecimal;
+
 @Getter
 @AllArgsConstructor
 public class ClientDto {
     private Long id;
     private String mail;
     private String address;
-    private Integer balance;
+    private BigDecimal balance;
 
 }
